NLP2/requirements.txt
langchain
langchain-openai
langgraph
python-dotenv
pandas
matplotlib
